LayoutTests/fast/dom/MutationObserver/mutation-observer-constructor-expected.txt

Test the constructor of WebKitMutationObserver

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.WebKitMutationObserver is non-null.
PASS typeof WebKitMutationObserver.prototype.observe is "function"
PASS typeof WebKitMutationObserver.prototype.disconnect is "function"
PASS typeof observer.observe is "function"
PASS typeof observer.disconnect is "function"
PASS new MutationObserver({ handleEvent: function() {} }) threw exception TypeError: Callback argument must be a function.
PASS new MutationObserver({}) threw exception TypeError: Callback argument must be a function.
PASS new MutationObserver(42) threw exception TypeError: Callback argument must be a function.
PASS new MutationObserver("foo") threw exception TypeError: Callback argument must be a function.
PASS successfullyParsed is true

TEST COMPLETE
